+++ changelog.md
@@ -10,6 +10,7 @@
 ### Patch
 
 - Fixed class exports not being detected, via [#11](https://github.com/jaydenseric/find-unused-exports/pull/11).
+- Fixed dynamic imports within default or named exports not being detected.
 - Updated dev dependencies.
 - Updated package scripts.
 - Updated GitHub Actions CI config: